refactor: delete BidiViewModel to reduce abstraction
