+++ firebase-ai/app/src/main/java/com/google/firebase/quickstart/ai/feature/live/StreamAudioViewModel.kt
@@ -1,5 +1,7 @@
 package com.google.firebase.quickstart.ai.feature.live
 
+import android.annotation.SuppressLint
+import androidx.lifecycle.ViewModel
 import com.google.firebase.Firebase
 import com.google.firebase.ai.ai
 import com.google.firebase.ai.type.AudioTranscriptionConfig
@@ -7,14 +9,20 @@ import com.google.firebase.ai.type.FunctionCallPart
 import com.google.firebase.ai.type.FunctionDeclaration
 import com.google.firebase.ai.type.FunctionResponsePart
 import com.google.firebase.ai.type.GenerativeBackend
+import com.google.firebase.ai.type.LiveSession
 import com.google.firebase.ai.type.PublicPreviewAPI
 import com.google.firebase.ai.type.ResponseModality
 import com.google.firebase.ai.type.Schema
 import com.google.firebase.ai.type.SpeechConfig
 import com.google.firebase.ai.type.Tool
+import com.google.firebase.ai.type.Transcription
 import com.google.firebase.ai.type.Voice
 import com.google.firebase.ai.type.liveGenerationConfig
 import com.google.firebase.quickstart.ai.feature.text.functioncalling.WeatherRepository.Companion.fetchWeather
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonObject
@@ -23,8 +31,23 @@ import kotlinx.serialization.json.jsonPrimitive
 @Serializable
 object StreamRealtimeAudioRoute
 
+enum class TranscriptionSpeaker {
+    USER,
+    MODEL
+}
+
+data class TranscriptionItem(
+    val speaker: TranscriptionSpeaker,
+    val text: String
+)
+
 @OptIn(PublicPreviewAPI::class)
-class StreamAudioViewModel : BidiViewModel() {
+class StreamAudioViewModel : ViewModel() {
+    private var liveSession: LiveSession
+
+    private val _transcriptions = MutableStateFlow<List<TranscriptionItem>>(emptyList())
+    val transcriptions: StateFlow<List<TranscriptionItem>> = _transcriptions.asStateFlow()
+
     init {
         val liveGenerationConfig = liveGenerationConfig {
             speechConfig = SpeechConfig(voice = Voice("CHARON"))
@@ -63,7 +86,7 @@ class StreamAudioViewModel : BidiViewModel() {
         runBlocking { liveSession = liveModel.connect() }
     }
 
-    override fun handler(functionCall: FunctionCallPart): FunctionResponsePart {
+    fun handleFunctionCall(functionCall: FunctionCallPart): FunctionResponsePart {
         val response: JsonObject
         if (functionCall.name == "fetchWeather") {
             val city = functionCall.args["city"]?.jsonPrimitive?.content
@@ -81,5 +104,46 @@ class StreamAudioViewModel : BidiViewModel() {
             response = JsonObject(emptyMap())
         }
         return FunctionResponsePart(functionCall.name, response, functionCall.id)
+    }
+
+    private fun handleTranscription(input: Transcription?, output: Transcription?) {
+        input?.text?.let { text ->
+            if (text.isNotEmpty()) {
+                _transcriptions.update { current ->
+                    val last = current.lastOrNull()
+                    if (last != null && last.speaker == TranscriptionSpeaker.USER) {
+                        current.dropLast(1) + last.copy(text = last.text + text)
+                    } else {
+                        current + TranscriptionItem(speaker = TranscriptionSpeaker.USER, text = text)
+                    }
+                }
+            }
+        }
+        output?.text?.let { text ->
+            if (text.isNotEmpty()) {
+                _transcriptions.update { current ->
+                    val last = current.lastOrNull()
+                    if (last != null && last.speaker == TranscriptionSpeaker.MODEL) {
+                        current.dropLast(1) + last.copy(text = last.text + text)
+                    } else {
+                        current + TranscriptionItem(speaker = TranscriptionSpeaker.MODEL, text = text)
+                    }
+                }
+            }
+        }
+    }
+
+    // The permission check is handled by the view that calls this function.
+    @SuppressLint("MissingPermission")
+    suspend fun startConversation() {
+        liveSession.startAudioConversation(::handleFunctionCall, ::handleTranscription)
+    }
+
+    fun endConversation() {
+        liveSession.stopAudioConversation()
+    }
+
+    fun clearTranscriptions() {
+        _transcriptions.value = emptyList()
     }
 }
